+++ src/main/kotlin/README.md
@@ -13,7 +13,7 @@
   - 숫자 하나당 4 종류의 카드
 - [X] 게임에 참여할 사람 입력 받기
   - 쉼표 기준으로 분리하기
-- [ ] 무작위로 2장 나눠주기
+- [X] 무작위로 2장 나눠주기
 - [ ] 입력받은 순서대로 `n` 입력을 받을 때까지 카드 계속 받기
   - ACE 점수 고려하기 (1 or 11)
 - [ ] 결과 출력하기
